refactor: improve component folder structure
